Update VkontakteManager.m
There is a runtime error happening only in last RN versions (might be new iOS version and Obj-C). The error is "Terminating app due to uncaught exception 'NSInvalidArgumentException', reason: '-[NSTaggedPointerString stringValue]: unrecognized selector sent to instance 0xbb689ad41de6145a'". So following changes work for me. It is not a lot but I thought it would cool to open a PR.